refactoring to remove the complex structure of recipient control
