+++ components/Conversation/RecipientControl.tsx
@@ -4,11 +4,7 @@ import AddressInput from '../AddressInput'
 import { checkIfPathIsEns } from '../../helpers'
 import { useAppStore } from '../../store/app'
 import useWalletProvider from '../../hooks/useWalletProvider'
-
-type RecipientInputProps = {
-  recipientWalletAddress: string | undefined
-  onSubmit: (address: string) => Promise<void>
-}
+import BackArrow from '../BackArrow'
 
 const RecipientInputMode = {
   InvalidEntry: 0,
@@ -18,13 +14,11 @@ const RecipientInputMode = {
   NotOnNetwork: 4,
 }
 
-const RecipientControl = ({
-  recipientWalletAddress,
-  onSubmit,
-}: RecipientInputProps): JSX.Element => {
+const RecipientControl = (): JSX.Element => {
   const { resolveName, lookupAddress } = useWalletProvider()
   const client = useAppStore((state) => state.client)
   const router = useRouter()
+  const recipientWalletAddress = router.query.recipientWalletAddr as string
   const [recipientInputMode, setRecipientInputMode] = useState(
     RecipientInputMode.InvalidEntry
   )
@@ -34,21 +28,26 @@ const RecipientControl = ({
     async (address: string): Promise<boolean> => {
       return client?.canMessage(address) || false
     },
-    []
+    [client]
   )
 
-  const completeSubmit = useCallback(
-    async (address: string, input: HTMLInputElement) => {
-      if (await checkIfOnNetwork(address)) {
-        onSubmit(address)
-        input.blur()
-        setRecipientInputMode(RecipientInputMode.Submitted)
-      } else {
-        setRecipientInputMode(RecipientInputMode.NotOnNetwork)
-      }
-    },
-    [checkIfOnNetwork, onSubmit]
-  )
+  const onSubmit = async (address: string) => {
+    router.push(address ? `/dm/${address}` : '/dm/')
+  }
+
+  const handleBackArrowClick = useCallback(() => {
+    router.push('/')
+  }, [router])
+
+  const completeSubmit = async (address: string, input: HTMLInputElement) => {
+    if (await checkIfOnNetwork(address)) {
+      onSubmit(address)
+      input.blur()
+      setRecipientInputMode(RecipientInputMode.Submitted)
+    } else {
+      setRecipientInputMode(RecipientInputMode.NotOnNetwork)
+    }
+  }
 
   useEffect(() => {
     const handleAddressLookup = async (address: string) => {
@@ -86,7 +85,7 @@ const RecipientControl = ({
         await completeSubmit(recipientValue, input)
       }
     },
-    [completeSubmit, resolveName]
+    [resolveName]
   )
 
   const handleInputChange = useCallback(
@@ -110,47 +109,52 @@ const RecipientControl = ({
   )
 
   return (
-    <div className="flex-1 flex-col shrink justify-center flex h-[72px] bg-zinc-50 md:border-b md:border-gray-200 md:px-4 md:pb-[2px]">
-      <form
-        className="w-full flex pl-2 md:pl-0 h-8 pt-1"
-        action="#"
-        method="GET"
-        onSubmit={handleSubmit}
-      >
-        <label htmlFor="recipient-field" className="sr-only">
-          Recipient
-        </label>
-        <div className="relative w-full text-n-300 focus-within:text-n-600">
-          <div className="absolute top-1 left-0 flex items-center pointer-events-none text-md md:text-sm font-medium md:font-semibold">
-            To:
+    <>
+      <div className="md:hidden flex items-center ml-3">
+        <BackArrow onClick={handleBackArrowClick} />
+      </div>
+      <div className="flex-1 flex-col shrink justify-center flex h-[72px] bg-zinc-50 md:border-b md:border-gray-200 md:px-4 md:pb-[2px]">
+        <form
+          className="w-full flex pl-2 md:pl-0 h-8 pt-1"
+          action="#"
+          method="GET"
+          onSubmit={handleSubmit}
+        >
+          <label htmlFor="recipient-field" className="sr-only">
+            Recipient
+          </label>
+          <div className="relative w-full text-n-300 focus-within:text-n-600">
+            <div className="absolute top-1 left-0 flex items-center pointer-events-none text-md md:text-sm font-medium md:font-semibold">
+              To:
+            </div>
+            <AddressInput
+              recipientWalletAddress={recipientWalletAddress}
+              id="recipient-field"
+              className="block w-[95%] pl-7 pr-3 pt-[3px] md:pt-[2px] md:pt-[1px] bg-transparent caret-n-600 text-n-600 placeholder-gray-500 focus:outline-none focus:placeholder-gray-400 focus:ring-0 focus:border-transparent text-lg font-mono"
+              name="recipient"
+              onInputChange={handleInputChange}
+            />
+            <button type="submit" className="hidden" />
           </div>
-          <AddressInput
-            recipientWalletAddress={recipientWalletAddress}
-            id="recipient-field"
-            className="block w-[95%] pl-7 pr-3 pt-[3px] md:pt-[2px] md:pt-[1px] bg-transparent caret-n-600 text-n-600 placeholder-gray-500 focus:outline-none focus:placeholder-gray-400 focus:ring-0 focus:border-transparent text-lg font-mono"
-            name="recipient"
-            onInputChange={handleInputChange}
-          />
-          <button type="submit" className="hidden" />
-        </div>
-      </form>
+        </form>
 
-      {recipientInputMode === RecipientInputMode.Submitted ? (
-        <div className="text-md text-n-300 text-sm font-mono ml-10 md:ml-8 pb-1 md:pb-[1px]">
-          {hasName ? recipientWalletAddress : <br />}
-        </div>
-      ) : (
-        <div className="text-sm md:text-xs text-n-300 ml-[29px] pl-2 md:pl-0 pb-1 md:pb-[3px]">
-          {recipientInputMode === RecipientInputMode.NotOnNetwork &&
-            'Recipient is not on the XMTP network'}
-          {recipientInputMode === RecipientInputMode.FindingEntry &&
-            'Finding ENS domain...'}
-          {recipientInputMode === RecipientInputMode.InvalidEntry &&
-            'Please enter a valid wallet address'}
-          {recipientInputMode === RecipientInputMode.ValidEntry && <br />}
-        </div>
-      )}
-    </div>
+        {recipientInputMode === RecipientInputMode.Submitted ? (
+          <div className="text-md text-n-300 text-sm font-mono ml-10 md:ml-8 pb-1 md:pb-[1px]">
+            {hasName ? recipientWalletAddress : <br />}
+          </div>
+        ) : (
+          <div className="text-sm md:text-xs text-n-300 ml-[29px] pl-2 md:pl-0 pb-1 md:pb-[3px]">
+            {recipientInputMode === RecipientInputMode.NotOnNetwork &&
+              'Recipient is not on the XMTP network'}
+            {recipientInputMode === RecipientInputMode.FindingEntry &&
+              'Finding ENS domain...'}
+            {recipientInputMode === RecipientInputMode.InvalidEntry &&
+              'Please enter a valid wallet address'}
+            {recipientInputMode === RecipientInputMode.ValidEntry && <br />}
+          </div>
+        )}
+      </div>
+    </>
   )
 }
